requirements.txt
Flask
python-dotenv
Flask-SQLAlchemy
psycopg2-binary
Flask-Cors
PyJWT
